Added class
