+++ .vscode/settings.json
@@ -1,0 +1,3 @@
+{
+    "python.pythonPath": "C:\\Users\\henry\\anaconda3\\envs\\internship_repo\\python.exe"
+}
