+++ datascript_class.py
@@ -3,27 +3,31 @@ from urllib.request import urlretrieve as retrieve
 import zipfile
 # import pandas as pd
 # import numpy as np
-import os
-import shutil
+# import os
+# import shutil
 
-class GenerationData(object):
-    def __init__(self, user=None, msg=None):
-        self.url = 'https://www.eia.gov'
-        self.eia_2018_location = '/electricity/data/eia860/archive/xls/eia8602018.zip'
-        self.user = user
-        self.print_user(msg)
-        return
+class GetData(object):
+    def __init__(self, page_url='https://www.eia.gov/electricity/data/eia860/'):
+        self.page_url = page_url
+        self.eia860_location = 'archive/xls/eia860'
 
-    def print_user(self, msg=''):
-        print(f'{msg} {self.user}')
-        return
 
     # download zip from website
-    def get_zip(self,url, zip_file_name):
-        retrieve(url, zip_file_name)
+    def get_zip(self, year, zip_file_name):
+        retrieve(f'{self.page_url}{self.eia860_location}{year}.zip', f'{zip_file_name}_{year}')
 
     # extract zip
     def extract_zip(self, zip_to_be_extracted, file_name):
         with zipfile.ZipFile(zip_to_be_extracted, 'r') as zip_ref:
             zip_ref.extractall(file_name)
 
+
+    # def get_data(self, page_url, year): 
+
+'''
+https://www.eia.gov/electricity/data/eia860/archive/xls/eia8602018.zip
+https://www.eia.gov/electricity/data/eia860/xls/eia8602019.zip
+https://www.eia.gov/electricity/data/eia860/archive/xls/eia8602017.zip
+https://www.eia.gov/electricity/data/eia860/archive/xls/eia8602015.zip
+https://www.eia.gov/electricity/data/eia860/archive/xls/eia8602011.zip
+'''
